Updates to printed messages and unit tests
[1] The messages printed by the program and some function names were
edited to make it clearer that the program is designed to *convert*
ISA-TAB-Nano datasets based on Excel files to datasets based on
TAB-delimited text files, including fixing some potential typos. It is
*not* a "parser".

[2] Some additional unit tests were added and some typos in existing
unit tests fixed to make sure that aspects of the code's functionality
were better checked.

[3] These changes were not intended to, and do not appear to have, cause
any change in the functionality of this program.